Add scanner.go
